+++ Assets/Scripts/Enums/MoveType.cs
@@ -4,7 +4,7 @@ namespace Assets.Scripts.Enums
     /// <summary>
     /// an Enum representing the Move type , whether it's capturing , Normal move , Check ...
     /// </summary>
-    public enum MoveType : ushort
+    public enum MoveType : byte
     {
         None = 0,
         Normal,
